Unmaintained status
Open source projects abusing actively developed APIs are a nightmare to maintain, especially when it competes with commercial offerings. Over the course of 6 months, constantly fighting with OpenAI has been fun. However, it is by no means sustainable. They will constantly update and open source efforts are easily overturned. Due to the ease by which OpenAI can shut our projects down, it has also lead to a lack of open contributions and a hoarding of knowledge. If anyone wishes to spend the time and effort to continue maintaining this project, feel free to fork. If a certain variant gains popularity or has shown great progress, I will put that on the top of the readme of this project to redirect users there.

Thank you to all the contributors so far (a whopping 122 individuals), @linweiyuan, and @pengzhile for keeping this project alive.